requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
bcrypt==4.0.1
certifi==2024.8.30
click==8.1.7
dill==0.3.9
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
lxml==5.3.0
mccabe==0.7.0
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
setuptools==75.2.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.31.1
